requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
--find-links https://data.pyg.org/whl/torch-1.13.0+cu117.html
torch==1.13.0+cu117
torch-geometric==2.3.1
torch-scatter==2.1.1
torch-sparse==0.6.17
torch-cluster==1.6.1
torch-spline-conv 
torchdrug==0.2.1
splito
umap-learn
scipy # pharmaconet
timm==0.9 # pharmaconet
omegaconf>=2.3 # pharmaconet
meeko
scipy
networkx
tensorboard
cvxopt
GitPython
numpy
botorch
biopython
gpytorch
easydict
pandas
tqdm
lmdb 
wandb
torchmetrics
CairoSVG
omegaconf>=2.3
git+https://github.com/dptech-corp/Uni-Dock.git#egg=unidock_tools&subdirectory=unidock_tools
